requirements.txt
discord.py
python-dotenv
yfpy